+++ src/vidore_benchmark/retrievers/colqwen2_retriever.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 import logging
+import os
 from typing import List, Optional, Union, cast
 
 import torch
@@ -28,8 +29,9 @@ class ColQwen2Retriever(VisionRetriever):
 
     def __init__(
         self,
-        pretrained_model_name_or_path: str = "vidore/colqwen2-v0.1",
+        pretrained_model_name_or_path: str = "vidore/colqwen2-v1.0",
         device: str = "auto",
+        num_workers: Optional[int] = None,
     ):
         super().__init__()
 
@@ -62,6 +64,14 @@ class ColQwen2Retriever(VisionRetriever):
         )
         print("Loaded custom processor.\n")
 
+        if num_workers is None:
+            if self.device == "mps":
+                self.num_workers = 0  # MPS does not support dataloader multiprocessing
+            else:
+                self.num_workers = os.cpu_count() if os.cpu_count() is not None else 1
+        else:
+            self.num_workers = num_workers
+
     @property
     def use_visual_embedding(self) -> bool:
         return True
@@ -78,6 +88,7 @@ class ColQwen2Retriever(VisionRetriever):
             batch_size=batch_size,
             shuffle=False,
             collate_fn=self.process_queries,
+            num_workers=self.num_workers,
         )
 
         query_embeddings: List[torch.Tensor] = []
@@ -95,6 +106,7 @@ class ColQwen2Retriever(VisionRetriever):
             batch_size=batch_size,
             shuffle=False,
             collate_fn=self.process_images,
+            num_workers=self.num_workers,
         )
 
         passage_embeddings: List[torch.Tensor] = []